criando trativas de erros e criando a controller do ingredintes

+++ src/services/DishesCreateService.js
@@ -1,4 +1,5 @@
 const DiskStorage = require("../providers/DiskStorage");
+const AppError = require("../utils/AppError");
 
 class DishesCreateService {
     constructor(dishesRepository) {
@@ -13,41 +14,54 @@ class DishesCreateService {
     }
 
     async show({ id }) {
-        const dishes = await this.dishesRepository.findById({ id });
-        return dishes;
+        try {
+            const dishes = await this.dishesRepository.findById({ id });
+            return dishes;
+        } catch (error) {
+            throw new AppError("Erro ao buscar prato", 404);
+        }
     }
 
     async execute({ img, name, category_id, ingredients, price, description }) {
-        const diskStorage = new DiskStorage();
-        const fileName = await diskStorage.saveFile(img);
-
-        const dishesCreated = await this.dishesRepository.createDishes({ img: fileName, name, category_id, ingredients, price, description });
-
-        return dishesCreated;
+        try {
+            const diskStorage = new DiskStorage();
+            const fileName = await diskStorage.saveFile(img);
+            const dishesCreated = await this.dishesRepository.createDishes({ img: fileName, name, category_id, ingredients, price, description });
+            return dishesCreated;
+        } catch (error) {
+            throw new AppError("Erro ao criar prato", 400);
+        }
     }
 
     async update({ id, img, name, category_id, ingredients, price, description }) {
-        const diskStorage = new DiskStorage();
+        try {
+            const diskStorage = new DiskStorage();
+            const dishes = await this.dishesRepository.findById({ id });
+            let fileName;
 
-        const dishes = await this.dishesRepository.findById({ id });
-        let fileName;
-
-        if (img != null || img != undefined || img != '' || "") {
-            if (dishes.img) {
-                await diskStorage.deleteFile(dishes.img);
+            if (img != null || img != undefined || img != '' || "") {
+                if (dishes.img) {
+                    await diskStorage.deleteFile(dishes.img);
+                }
+                fileName = await diskStorage.saveFile(img);
+            } else {
+                fileName = dishes.img;
             }
-            fileName = await diskStorage.saveFile(img);
-        } else {
-            fileName = dishes.img;
-        }
-        const dishesUpdate = await this.dishesRepository.updateDishes({ id, img: fileName, name, category_id, ingredients, price, description });
 
-        return dishesUpdate;
+            const dishesUpdate = await this.dishesRepository.updateDishes({ id, img: fileName, name, category_id, ingredients, price, description });
+
+            return dishesUpdate;
+        } catch (error) {
+            throw new Error("Erro ao atualizar prato: ", 500);
+        }
     }
 
     async delete({ id }) {
         const diskStorage = new DiskStorage();
         const dishes = await this.dishesRepository.findById({ id });
+        if (!dishes) {
+            throw new AppError("Prato não encontrado", 404);
+        }
 
         if (dishes.img) {
             await diskStorage.deleteFile(dishes.img);
